+++ src/Assets/CustomAttributesAssetBundle.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Hirtz\Skeleton\Assets;
+
+class CustomAttributesAssetBundle extends AbstractAssetBundle
+{
+    public $js = ['js/custom-attributes.js'];
+}